Remove redundant links
The download links for the Sysinternals Suite appear twice in this short
page. The ones at the end of the page seems to not be maintained (in
terms of updating the file sizes), so it seems most people see and use
the top set of links.
In light of the above, and to improve overall maintainability, removing
the set of links at the bottom of the page.